Use the new macros in the main interface
